+++ examples/MultiChannel/MultiChannelProcessor.h
@@ -14,7 +14,7 @@ public:
     ~MultiChannelProcessor(){};
 
     // Process method to handle audio processing
-    void process(double** inputs, double** outputs, std::size_t blockSize) override;
+    void process(float** inputs, float** outputs, std::size_t blockSize) override;
 
     // Set a parameter value
     void setParameter(int paramId, double value) override;
